requirements.txt
torch
segmentation_models_pytorch
torchvision
tqdm
numpy
albumentations
PyYAML
torchsummary
scikit-learn
opencv-python
tensorboard